Implement basic login timeout support via a 'loginTimeout' URL parameter / property. Also supported via setLoginTimeout on the DataSource variants.
The approach I've taken is that when a login timeout is specified, the
actual connection attempt happens in a new thread. The original caller
blocks waiting for a result from that thread, and can time out as desired.

The driver will temporarily leak threads when timeouts actually occur.
The new connect thread always waits for the connection to complete or
fail, and won't exit until that happens.

There is one case where we could permanently leak threads: when the login
process blocks indefinitely (for example, if the specified port accepts
the connection and then does not respond to the startup packet at all).
To fix this requires more invasive changes than I'm willing to do at the
moment. We would need to be able to get at the protocol stream to close
it before the connection is fully established, but currently all the
connection logic is hidden in the JdbcNConnection ctor.

If no login timeout is specified, connections are established in the
normal way with no extra threads involved.