The latest javadocs have added some clarity to what the precision of non-numeric fields should be.  For example a date/time type's precision is the length of the String representation.
It also seems to hint that we shouldn't be returning -1 for unknown
lengths and since we've already seen RowSet implementations assume
this go along and return 0 instead.

Finally correct a misunderstanding about the varbit type.  I assumed
that since you couldn't say varbit(N) it didn't have a valid typmod,
but the correct syntax is bit varying(N) so account for typmod
correctly.