Change the ResultSet to cache ResultSetMetaData instances.
This dramatically improves performance with regards to
metadata-intensive operations. The original motivation is that without
caching, each call to a ResultSet.updateXXX method would retrieve a new
instance of the metadata, causing an entire query to be run.  If you
update many columns in the same row, the cost came to be very expensive.

Performance may improve in other workloads as well, but this has not been
measured.