Move logic out of concrete JDBC version classes.
The concrete JDBC implementation classes should do as little work
as possible.  They exist solely to connect the right set of concrete
implemention classes, leaving all the real work to the Abstract
versions.  Some logic had started to creep out into these concrete
classes which is bad because it duplicates code in paths that aren't
likely to be tested by a developer who is working primarily with a
single JDK version.