Add Vagrant VM for testing and SSL tests for SingleCertSocketFactory
This patch adds a Vagrant configuration for testing the PG JDBC driver.
Once you've installed Vagrant (http://www.vagrantup.com/) bring up the VM by running:

    $ vagrant up

This will bring up an Ubuntu 12.04 VM and then:

* Runs a software update
* Installs PostgresQL
* Changes the listen address to '*' in postgresql.conf
* Create a couple of test databases and users
* Modify pg_hba.conf to allow remote connections to the test databases/users
* Restart the PostgreSQL daemon (so the changes take effect)
* Forward the VM's PostgreSQL server port of 5432 to 15432 on the local machine

The first time the VM is created it will take a few minutes to update and install everything.
After that the VM is configured to only check for updates on startup once a week.
The initial database and schema creation only happens when the VM is first created.
Modifications to the database (or VM itself) will be persisted until it is destroyed.

Once testing is complete, to shutdown the VM:

    $ vagrant halt

The ssl-test directory includes a Maven project to test the SingleCertSocketFactory.
The tests are configured to run against the Vagrant VM's PostgreSQL server.
To run the tests build the driver as normal (e.g. "ant jar") then change directory to the ssl-test directory and run:

    $ mvn clean test

See the README.md in the ssl-test directory for more details on running the tests.