Resolving Duplicate Column Names
The JDBC spec says when you have duplicate columns names, the first one
should be returned but this behavior isn't useful to us. What we really
need is the ability to resolve those ambiguous cases in the client code.

I propose adding an additional string  to the columnNameIndexMap that
would be in the format of tablename.column similar to how you would
reference the columns in SQL.

I've added a test case for this use case that shows that it works and
doesn't affect any existing test cases.