Add an assumeMinServerVersion connection property.
Add an assumeMinServerVersion connection property to specify the expected server_version of the remote server.
If this property is set to 9.0+ (ex: 9.0, 9.1, 9.2, 9.3, etc) then the driver will add additional parameters to the StartupMessage that would otherwise be sent afterward in run InitialQueries().
For 9.0+ database this will save a number of round trips at connection creation as the parameters are otherwise set via individual SET statements in runInitialQueries().