Don't prevented VM shutdown if ConnectThread is blocked
ConnectThread prevented VM shutdown if the underlying socket operation was blocked and could not be interrupted.
Making the ConnectThread a daemon will allow a proper shutdown. 
The VM won't exit too early, as the Thread waiting for ConnectThread until the login timeout will keep the VM alive (because it shouldn't be a daemon).