Add SSPI support to PgJDBC
This change will cause code that was previously attempting direct SSPI
negotiation against a Kerberos server to instead use Windows' SSPI support
if it is available.

SSPI support using waffle-jni is always compiled, but the libraries are
not included in the default PgJDBC jar. They must be added on the classpath
or included in a "fat jar" for SSPI to work. When missing, PgJDBC throws
an informative exception.

This commit changes the default behaviour of the driver to use SSPI, but only if:

* Waffle-JNA is present on the classpath;
* We're running on Windows; and
* 'sspi' mode was requested by the server.

... so unless a project already has Waffle-JNA on its classpath for other
reasons you won't notice the change unless you take explicit action.

The new parameter `gsslib`, modeled after the libpq parameter of the same
name, is added to allow users to override whether SSPI is used where it
is available. The default is automatic selection as outlined above.

This parameter also lets you force use of SSPI against a server that makes a
`gssapi` mode auth request. That way a Windows client with existing
single-sign-on credentials for an spnego-enabled Kerberos server can re-use its
existing credentials via SSPI. Unlike libpq, which defaults to using SSPI on
Windows, PgJDBC will not attempt to use SSPI for a GSSAPI auth request this
unless explicitly forced to, as most Java users will prefer to do GSSAPI via
JSSE rather than via Windows SSPI.