Fail-fast if a resultset is returned but its structure is unknown
If a resultset gets returned but we don't have describe data for it and
don't know how the field representation should look, fail immediately.
Without this change the failure will occur much later, when the client
code attempts to access the result set.

This shouldn't get triggered under normal circumstances and indicates
a protocol violation or internal PgJDBC bug.