Add a new test case demonstrating a bug in returning support
If a prepared statement created with

    Connection.prepareStatement(String, String[])

to return values via getGeneratedKeys() (internally via RETURNING) is
executed in a batch, and the arguments to the batch include a string
value assigned via setObject that's null in the first execution but not
subsequent ones, PgJDBC's support for returning generated keys breaks.

The conditions for this to occur are odd to say the least, as
outlined in more detail in this test case.