Disable binary xfer on batches returning generated keys
If a batch that returns generated keys is re-planned mid-execution
because of an input parameter type change, we discard our knowledge
of the field types as obtained at pre-describe time.

This might change a decision about whether to request a result as
binary. If the result was something we'd usually send as as binary
(e.g. int4) then we might request a binary result for the first few
executions. Then if the query gets reparsed and replanned midway
through the batch we'll throw away our knowledge of the fields. Even if
we immediately Describe again, we don't get that result until later, so
we'll still be working with unknown fields when doing parameter binding.
If we don't know the fields we don't specify the result format
requested, so the server sends text.

This means we'll get a mixture of text and binary results for the same
column in the result set, causing errors when we try to decode the
result set row buffer.

There's really no way for us to know in advance whether we can support
binary mode for the result types of every execution, since the result
types can be determined by the input parameters. So just turn it off.

This can happen even if the user doesn't think they're changing the
type. Take the case explained in issue #266, where:

    setObject(1, null)

infers type 'text' then a later execution:

    setObject(1, "text")

sets type 'varchar' on the driver side. That'll cause a re-plan.

(Arguably we should be retaining our knowledge of the Statement level
describe, not throwing it away on re-plan, but that's a separate patch
to consider).