perf: Remove AbstractJdbc2Statement.finalize()
Removes AbstractJdbc2Statement.finalize() as server side resource
cleanup is already handled byQueryExecutorImpl#processDeadParsedQueries.

With this change, if many prepared statements are created, but not
closed, and no further activity occurs on the connection, then the
server side resources for those connections will never be closed. The
scenario is considered to be unlikely and in any event is incorrect
usage of the driver.