bug: equals on an array does not return true when values are equal
All fields returned by the JDBC driver are expected to return true with
the equals function when the values are the same.  This commit
overrides the equals function on the array object so that the state is
compared as opposed to just the identity.

It is assumed that both fieldBytes and fieldString cannot be null at
the same time.  Breaking that assumption would require that a null is
passed into the constructor.