feat: Improved composite/array type support and type naming changes.
Composite types can live if their own schema. This requires type name lookup to cope with it.
Names are case sensitive.
Added simple logic which will split schema/name when . is used.
Postgres allows for names which start with '_'; therefore it is incorrect to assume array type will start with '_', although that can be assumed for all builtin types.
Improve type detection and use '[]' suffix as array alias.
Make most of the code backward compatible so it uses old logic with '_' instead of '[]' on lookups and naming.
Relations which live in search_path will be registered using both simple and full name.
Simple arrays will be populated using connection.getObject if everything else fails (only 1-dim array are handled as such)

TODO: handle scenario when '"' is used inside schema or type name.