Deadlock after IO exception during copy cancel
During a copy, the query executor lock is taken and held until the copy
either completes successfully or is cancelled.  In the latter case, if
an IOException occurs during the cancel the lock can be left unreleased.
Afterward, no operation on the connection (e.g., a rollback in a finally
block) can even fail because it will wait for the lock to be released.