Allow both single and double question marks in simple statements
This is an improvement over the previous commit:
Now you can either use a single question mark in a java.sql.Statement
which will be interpreted as such, or you can use escaped (doubled)
question marks.
This way code that used to work will continue to work, while people
who do not expect that question marks need to be escaped in simple
statements will not be surprised.

This commit also fixes the regression tests and adds a new one.