feat: Support for Simple queries on v3 protocol.
Postgres supports both simple and extended queries on v3 protocol.
Prepared statementes are always executed with extended protocol.
This commit introduces option to change the behavior of non-prepared statements, so they are not parsed on the client side and are sent to the server as simple queries.
This fixes the problem of large SQL statements hanging, since it sends them with a known size.
By default, large SQL statement are parsed into subqueries and sent to the server as large collection of Parse/Bind/Execute commands.
In some scenarios, this can lead to hanging of SQL statement.

Simple protocol format is:
Q
length
sql command text
\0

To enable this behavior set enableSimpleQueries=true connection parameter.