fix: Do not swallow security exceptions
There are two cases where the driver currently swallows security
exceptions. The first is in MD5Digest and the second is in
SingleCertValidatingFactory. In both cases the code will result in
partially initialized objects being used which will result in follow up
exceptions later on. These will be hard to track down since the
original exception was swallowed.

Wrap and propagate security exceptions instead of swallowing them.