fix: support local date times not in time zone
Local date time support introduced in #474 has subtle edge cases. These
occur when trying to read a LocalDateTime that does not exist in the VM
default time zone because it falls in a "not existing" time during a
daylight saving time offset transition. For LocalDateTime instances
this should not be an issue since they are explicitly not associated
with a time zone. This commit tests all time zones and timestamps from
TimezoneTest except 2015-06-30T23:59:60, this is only valid with a time
zone in order to know whether it is actually a leap second.

Fix local date time edge cases introduced by #474.
Closes #493