fix: NPE in DatabaseMetaData.getTypeInfo when types are dropped concurrently
DatabaseMetaData.getTypeInfo used several queries to load the data, thus the first of them
might observe some oid, then subsequent selects might fail to see it if the type was dropped.

The solution is not 100% bulletproof, however getTypeInfo always returns full info, thus it
is more likely to fail in a user-visible way.