javadoc: don't generate javadoc for not-compiled packages
This is unfortunate because users who want to build without OSGIi
and without Waffle need to specify yet another compilation flag.

But there is no way to do this automatically when user specifies
'-DwaffleEnabled', the reason is that excludePackageNames accepts
one string (colon or comma separated list of package names), and
we can not edit the excludePackageNames tag twice (firstly for
waffleEnabled and then for osgiEnabled properties) -- the last
excludePackageNames wins.