doc: fix brand names and links in readme (#636)
* Use correct capitalization and spacing for brand names in README

As of macOS Sierra the capitalization changed to have an initial
lowercase 'm' (and the X in OS X was dropped from the brandname).

* Move to using https for all postgresql.org links

The postgresql.org site is by default serving all traffic over https,
change links in README to avoid an extra redirect. The exception is
still developer.postgresql.org which only serves traffic over http.