fix: Don't break when there is no column metadata
It's possible to use Postgres in an object oriented way.
Driver shouldn't break since some of it's assumptions are not valid in those usages.

"SELECT t FROM table t" will return the whole object, instead of some column inside it.
In those cases FieldMetadata doesn't make too much sense, since it's not returning the field, but rather a whole tuple.

Changed the test not to break (removal of public.) since this is ommited when object is on search path.