feat: add support for PreparedStatement.setCharacterStream(int, Reader)
In pgjdbc 9.4 (1211) and earlier, the PreparedStatement method
setCharacterStream(int, Reader) throws a not implemented
exception. The setCharacterStream(int, Reader, int) method does not
throw an exception, but it materializes the Reader as a String and
then uses the same code path as setString(int, String). Materializing
the readers can cause memory usage issues (OutOfMemoryError) when
working with large text values.
This patch implements the setCharacterStream(int, Reader) method
using in memory buffers for small streams and temp files for large
streams. To do so, it uses code added in pull request #220 that was
used to implement setBinaryStream(int, InputStream). This patch
adds a utility class to convert the Reader's UTF-16 character stream
into a UTF-8 binary stream.
The setCharacterStream(int, Reader) method implemented in this patch
supports both simple and extended query modes. However, in simple
mode, it materializes the reader as a String just like
setCharacterStream(int, Reader, int).

fixes #671