test: check that new properties follow correct lower camel case
To avoid bad declaration of new properties in the driver, this test
check that the name of the enum matches the name of the property and that
for each "_" in the enum uses the lower camel case for the property.

There is an exclusion list for the properties that do not follow
the correct lower camel case but are retained for backward compatibility.

closes #727