fix: datasource can not get a connection if the driver is not registered
The current DataSource implementation relies in DriverManager to get a connection, but the DriverManager
requires that the driver is registered and when the org.postgresql.Driver is called it's automatically
registered in a static initializer block, but for the DataSource class it is not called.

To trigger the register of the driver, simply make a call to org.postgresql.Driver.isRegistered() which
checks if the driver is registered and initialize the class if not.

closes #768