fix: function calls without parameters do not require parentheses
According to the JDBC documentation, the parentheses surrounding
the parameters, of a function call, are optional if there are no
parameters. However, Parser required the parentheses to be present at
all times. The new behavior is that the Parser checks whether there are
parentheses, and if not simply adds them. Also, tests have been added
for both cases, function calls to functions with no parameters with and
without parentheses.

This allows to execute function calls like this:

    {?= call somefunction }

Which is line with the JDBC documentation.

The tracking bug is #767.