Improve finalizer guard mechanism to reduce GC overhead
The PgConnection class override finalize to guard against leaked connections
and be able to log them if they are leaked by the code.

While the goal is worthy, the implementation causes large overhead in Garbage
collection because, no matter if the leak guard is needed or not, the finalizer
will be executed.

This change introduces a finalizer guard following the pattern explained in the
"Effective Java" book, which uses a dedicated object with "just" the finalize
method overridden, and that it is instantiated only if the leak guard is needed.
If the leak guard is *not* needed, no finalizer will cause overhead during
Garbage Collection.