fix: cache *array* oid in _arrayOidToDelimiter
This previously didn't cause any obvious failures as it was populating
the cache with element type oids. The calling code does pass in arrray
oids, which resulted in a cache miss and a database lookup. In the
off-chance a caller were to pass in an element oid for a core type, it
would have wrongly hit the cache, rather than erroring out.

This fix also prevents the unnecessary database lookups for core types.