fix: respect search path for non-simple type names
Non-simple type names are quoted or an array. Previously, such type
names would match only if their schemas were on the search path, but
search path precedence was not respected.

For example, with types a.foo, b.foo, and c.foo, with search_path set to
'a,b', "foo" would match either a.foo or b.foo depending on which was
created most recently. c.foo would not be selected. This commit fixes
this behavior so a.foo will be selected as expected, regardless of
creation order.

Note that pg_catalog is always at the front of the search path;
user-defined types should never shadow system types. The non-simple
cases fixed in this patch *do* include pg_catalog at the front of the
search_path. While the simple case (not affected by this commit)
respects search path precendence for user schema (when set), it omits
pg_catalog from the search path allowing user-defined types to shadow
pg_catalog.