feat: parse command complete message via regex (#962)
Replaces command status whitelist based parsing with a regex approach to
handle generic COMMAND OID COUNT or COMMAND COUNT responses. If the
response does not match the regex then parsing is skipped. This should allow
for automatically supporting new server responses of that same form as well as
skipping any that cannot be parsed.

Fixes #958