fix: Added support for socksNonProxyHosts property (#975)
When a socksProxyHost is configured, there needs to be an escape value
so that some hosts are resolved immediately. The Java Networking spec[1]
does not specify such an option, but the socksNonProxyHosts is used in the
sun.net.spi.DefaultProxySelector. The behavior is mentioned in bug report
5001942[2]. This commit reimplements the logic in the DefaultProxySelector.select
method to determine if the host should be resolved immediately.

[1] http://docs.oracle.com/javase/8/docs/technotes/guides/net/proxies.html
[2] http://bugs.java.com/bugdatabase/view_bug.do?bug_id=5001942