The current MultiHostChooser implementation in the following two scenarios is not very reasonable.
1. jdbc:postgresql://slave1,master1/accounting?targetServerType=master
Each time the cached host state expires(control by hostRecheckSeconds), need to try 2 times to connect to the correct server.

improvement：
Adjust the connecting order, first try to connect master1
currently: slave1 -> master1
after    : master1 -> slave1

2. jdbc:postgresql://master1,slave1/accounting?targetServerType=preferSlave
The first connection or the first connection after cached host state expires, will be connected to the master1.
But that's not what we expected. Reasonable result is only in the absence of any available slave, will be connected to the Master.

improvement：
When targetServerType is preferSlave, internal attempts to make connection twice.
The first time internally set targetServerType to 'slave' and connect. If the connecting fails, then set the targetServerType to 'any' and do a second attempt.