fix: throw TOO_MANY_RESULTS (0100E) instead of "PgResultSet: tuples must be non-null"
When expecting no rows (e.g. executeUpdate with NO_GENERATED_KEYS) a query might still return one row
as "max rows" in Execute message does not allow to limit the resultset completely.

The returned tuples are ignored all the way up via "noResults / QueryExecutor.QUERY_NO_RESULTS", however
it fails when it meets handler.handleResultRows and tries to create a resultSet.

The corrective action is to create empty tuples list so empty resultSet is created.

The issue is reproduced in binary=FORCE in GeneratedKeysTest