fix: use 'time with time zone' and 'timestamp with time zone' values as is and avoid computation with user-provided/default Calendars
Previous behavior:
getTime(int), getTime(int, Calendar) kind of methods ensured that the resulting java.sql.Time
would render as 1970-01-01 ... in the given time zone (e.g. given Calendar and/or default TimeZone)
Note: the resulting Time object

New behavior:
for 'time with time zone' and 'timestamp with time zone' the value from the server is assumed to be
"instant" (i.e. absolute point in time), and no further adjustments is made to make the date part to be 1970-01-01

'time' and 'timestamp' work as earlier except "00:00:00" and "24:00:00" times in text format.
Previously, text times "00:00:00" and "24:00:00" were mapped to Time(0) and Time.valueOf("24:00:00"):
1) Time(0) is 1970-01-01 00:00:00 UTC (it does not account neither user-provided nor default Calendar)
2) Time.valueOf("24:00:00") uses system-provided calendar, and it does not account user-provided one