feat: Add cancelQuery() to PGConnection public interface
Adds cancelQuery() to the extended public interface PGConnection to allow
for issueing of cancels directly to connections rather than having to keep
track of individual statements. The internals of the driver and the backend
server already operate at a connection level (not a statement level) and
this patch only exposes the existing cancelQuery() method publicly.