fix: Encode url query parameters
BaseDataSource did not properly encode url parameters, meaning that users could
not log in if their password contained illegal characters. The bug can be
reproduced by setting the test user password to ';/?:@&=+$,' (a bunch of illegal
characters for query parameters). Encode the parameters. Strictly speaking the
parameter names should also be encoded but in this case they are a fixed set of
words which only consist of safe characters.

With the problem password, DriverTest also fails because it did not encode the
parameters either. Encode the parameters in the test too.