perf: remove unneeded cache memory allocations
When the cache is enabled through 'preferQueryMode', the driver
allocates more memory than if the cache is disabled
(transient allocations not the cache memory usage)

It comes from the cache key memory size estimation
(CachedQuery#getSize)
On cache borrow / release the size is estimated by generating a string
from the key and computing its length.

On simple tests it accounts for 60% of the allocated memory.

This patch creates a custom method in BaseQueryKey to estimate the size
without allocating too much memory.

Note that the new estimated size and the one before this patch might be
slightly different.