fix: encode url query parameters DataSource (#1201)
BaseDataSource did not properly encode url parameters, meaning that users could
not log in if their password contained illegal characters. The bug can be
reproduced by setting the test user password to ';/?:@&=+$,' (a bunch of illegal
characters for query parameters). Encode the parameters. Strictly speaking the
parameter names should also be encoded but in this case they are a fixed set of
words which only consist of safe characters.

With the problem password, DriverTest also fails because it did not encode the
parameters either. Encode the parameters in the test too.

* fix: do not leak password to URL

round-tripping a datasource through JNDI added the user and password
to the ds properties as well as to the instance variables - which was
not possible to do via setProperty. This may be a security issue if
the URL is logged, and was in part why passwords with non-url-safe
characters failed to connect in some circumstances.

Set properties using setProperty, so that consistent logic applies
to processing the user/password keys.