Re-use buffer and char array when parsing jdbc sql. Prior to this patch parsing a jdbc sql query was creating several buffers.
replaceProcessing (char array + buffer)
modifyJdbcCall if it's a callable statement (buffer)
parseJdbcSql (char array + buffer)

With this patch we re-user as much as possible the buffer and the char
array.

In replaceProcessing we track if any replacement was done.
If not (which is really common in orm based applications) we can re-use
the char array and we don't need to transform the buffer into a String.