postgresql-innotas-patch-42.1.4.2.6.jar
Provide InnotasCallback as a hook into PgResultSet from our app server, so we can tell it for what
tables to bypass getResultSetMetaData().getPrimaryKeys(), as that slows down inserting 1000 tasks for SSF.

Also, it now informs us whenever an insert, update, or delete statement is executed from an updatable result set,
as that SQL is never logged. Now, with PG_SQL_LOGGING turned on, that SQL is also logged.