test: Change drop object utilities in TestUtils to use IF EXISTS
Changes most drop object helper methods in TestUtils to check if a transaction
is in progress and, if not, then use IF EXISTS to suppress errors for missing
objects during the DROP command.

If a transaction is in progress then the command then the drop is executed
without the IF EXISTS clause so that missing objects results in an exception
and test failure.

In either case an unrelated error such as the server being offline will be now
throw an exception. This is a change from the prior behavior where all errors
that are not part of a transaction were being suppressed.
errors if the object being dropped does not exist.